+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and and th
 - Added a status bar item to execute all liquibase commands
 - Added an overview action in the status bar to create the db-doc of the database
 - Added a info message on first startup to inform the user about the existence of a walkthrough
+- Opened the output channel per default, if a command was executed. Behavior can be changed with the setting `openOutputChannelOnCommandStart`
 
 ### Changed